fix: avoid infinite redirect loop
When a client would visit /<municipality>/<admin-unit>/bogus/bogus2,
they would get trapped in an infinite redirect loop. This fixes that by
making sure the meeting id they're visiting exists before redirecting
(and otherwise 404-ing)
